**Breaking Changes**
- `MaildirPlusPlus` no longer inherits from `Maildir`. Use `getFolder("Inbox")` to get the inbox
  maildir.
- `Maildir::makeCurrent()` has been renamed to `touch()` and is now public.
- `$parentDir` has been changed to `$rootDir`.
- Deprecated function arguments are now dropped.

**Other changes**

- **Quotas support!** Added quota support functions to MaildirPlusPlus.
- Added isMaildir, rootDir, getPath, fetchAsStream and getStreams to Maildir.
- Maildir::createName is now public.

**Notes**

I decided to make MaildirPlusPlus no longer inherit from Maildir because I believe that
MaildirPlusPlus represents a collection of Maildirs rather than a Maildir in its own right. While
it could technichally masquerade as a Maildir; semantically, it doesn't really have an 'is_a'
relationship with Maildir. So I removed it out of a desire to engage in good programming practice.